Store firings with t timestamp & more data

+++ networks/Network.h
@@ -54,7 +54,7 @@ public:
   // Labels corresponding to the input
   std::vector<unsigned char> labels;
   // The timestamp, index and label of spikes stored
-  std::vector<std::tuple<int, int, int>> firings;
+  std::vector<std::tuple<float, int, int>> firings;
 
   cimg_library::CImg<unsigned char> im;
   cimg_library::CImgDisplay dis;
